front: move OperationalStudiesConfState to types.ts file
